requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.6.3
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
async-tio==1.3.2
attrs==22.1.0
biopython==1.81
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
curl-cffi==0.5.6
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.8.0
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.13.0
h5py==3.8.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.13.4
idna==3.3
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
OpenAIAuth==1.0.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
platformdirs==2.5.2
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
revChatGPT==6.1.4
rsa==4.9
scipy==1.10.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
sympy==1.11.1
telegram==0.0.1
termcolor==2.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tornado==6.1
tqdm==4.65.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.6
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
